[iOS] feat: AchievementRepository 통합
