+++ iOS/moti/moti/Domain/Sources/Domain/RepositoryProtocol/AchievementListRepositoryProtocol.swift
@@ -7,8 +7,20 @@
 
 import Foundation
 
-public protocol AchievementListRepositoryProtocol {
+public protocol AchievementRepositoryProtocol {
     func fetchAchievementList(
         requestValue: FetchAchievementListRequestValue?
     ) async throws -> ([Achievement], FetchAchievementListRequestValue?)
+    
+    func deleteAchievement(
+        requestValue: DeleteAchievementRequestValue
+    ) async throws -> Bool
+    
+    func fetchDetailAchievement(
+        requestValue: FetchDetailAchievementRequestValue
+    ) async throws -> Achievement
+    
+    func updateAchievement(
+        requestValue: UpdateAchievementRequestValue
+    ) async throws -> Bool
 }
